Make physic stateless
