+++ lib/physics.js
@@ -1,0 +1,476 @@
+const Vec3 = require('vec3').Vec3
+const AABB = require('./aabb')
+const math = require('./math')
+
+function Physics (mcData, world) {
+  const blocksByName = mcData.blocksByName
+
+  // Block Slipperiness
+  // https://www.mcpk.wiki/w/index.php?title=Slipperiness
+  const blockSlipperiness = {}
+  const slimeBlockId = blocksByName.slime_block ? blocksByName.slime_block.id : blocksByName.slime.id
+  blockSlipperiness[slimeBlockId] = 0.8
+  blockSlipperiness[blocksByName.ice.id] = 0.98
+  blockSlipperiness[blocksByName.packed_ice.id] = 0.98
+  if (blocksByName.frosted_ice) { // 1.9+
+    blockSlipperiness[blocksByName.frosted_ice.id] = 0.98
+  }
+  if (blocksByName.blue_ice) { // 1.13+
+    blockSlipperiness[blocksByName.blue_ice.id] = 0.989
+  }
+
+  // Block ids
+  const soulsandId = blocksByName.soul_sand.id
+  const webId = blocksByName.cobweb ? blocksByName.cobweb.id : blocksByName.web.id
+  const waterId = blocksByName.water.id
+  const lavaId = blocksByName.lava.id
+  const ladderId = blocksByName.ladder.id
+  const vineId = blocksByName.vine.id
+
+  const physics = {
+    gravity: 0.08, // blocks/tick^2 https://minecraft.gamepedia.com/Entity#Motion_of_entities
+    yawSpeed: 3.0,
+    sprintSpeed: 1.3,
+    stepHeight: 0.6, // how much height can the bot step on without jump
+    negligeableVelocity: 0.003 // actually 0.005 for 1.8, but seems fine
+  }
+
+  function getPlayerBB (pos) {
+    return new AABB(-0.3, 0, -0.3, 0.3, 1.8, 0.3).offset(pos.x, pos.y, pos.z)
+  }
+
+  function setPositionToBB (bb, pos) {
+    pos.x = bb.minX + 0.3
+    pos.y = bb.minY
+    pos.z = bb.minZ + 0.3
+  }
+
+  function getSurroundingBBs (world, queryBB) {
+    const surroundingBBs = []
+    const cursor = new Vec3(0, 0, 0)
+    for (cursor.y = Math.floor(queryBB.minY); cursor.y <= Math.floor(queryBB.maxY); cursor.y++) {
+      for (cursor.z = Math.floor(queryBB.minZ); cursor.z <= Math.floor(queryBB.maxZ); cursor.z++) {
+        for (cursor.x = Math.floor(queryBB.minX); cursor.x <= Math.floor(queryBB.maxX); cursor.x++) {
+          const block = world.getBlock(cursor)
+          if (block) {
+            const blockPos = block.position
+            for (const shape of block.shapes) {
+              const blockBB = new AABB(shape[0], shape[1], shape[2], shape[3], shape[4], shape[5])
+              blockBB.offset(blockPos.x, blockPos.y, blockPos.z)
+              surroundingBBs.push(blockBB)
+            }
+          }
+        }
+      }
+    }
+    return surroundingBBs
+  }
+
+  physics.adjustPositionHeight = (pos) => {
+    const playerBB = getPlayerBB(pos)
+    const queryBB = playerBB.clone().extend(0, -1, 0)
+    const surroundingBBs = getSurroundingBBs(world, queryBB)
+
+    let dy = -1
+    for (const blockBB of surroundingBBs) {
+      dy = blockBB.computeOffsetY(playerBB, dy)
+    }
+    pos.y += dy
+  }
+
+  function moveEntity (entity, world, dx, dy, dz) {
+    const vel = entity.vel
+    const pos = entity.pos
+
+    if (entity.isInWeb) {
+      dx *= 0.25
+      dy *= 0.05
+      dz *= 0.25
+      vel.x = 0
+      vel.y = 0
+      vel.z = 0
+      entity.isInWeb = false
+    }
+
+    const oldVelX = dx
+    const oldVelY = dy
+    const oldVelZ = dz
+
+    let playerBB = getPlayerBB(pos)
+    const queryBB = playerBB.clone().extend(dx, dy, dz)
+    const surroundingBBs = getSurroundingBBs(world, queryBB)
+    const oldBB = playerBB.clone()
+
+    for (const blockBB of surroundingBBs) {
+      dy = blockBB.computeOffsetY(playerBB, dy)
+    }
+    playerBB.offset(0, dy, 0)
+
+    for (const blockBB of surroundingBBs) {
+      dx = blockBB.computeOffsetX(playerBB, dx)
+    }
+    playerBB.offset(dx, 0, 0)
+
+    for (const blockBB of surroundingBBs) {
+      dz = blockBB.computeOffsetZ(playerBB, dz)
+    }
+    playerBB.offset(0, 0, dz)
+
+    // Step on block if height < stepHeight
+    if (physics.stepHeight > 0 &&
+        (entity.onGround || (dy !== oldVelY && oldVelY < 0)) &&
+        (dx !== oldVelX || dz !== oldVelZ)) {
+      const oldVelXCol = dx
+      const oldVelYCol = dy
+      const oldVelZCol = dz
+      const oldBBCol = playerBB.clone()
+
+      dy = physics.stepHeight
+      const queryBB = oldBB.clone().extend(oldVelX, dy, oldVelZ)
+      const surroundingBBs = getSurroundingBBs(world, queryBB)
+
+      const BB1 = oldBB.clone()
+      const BB2 = oldBB.clone()
+      const BB_XZ = BB1.clone().extend(dx, 0, dz)
+
+      let dy1 = dy
+      let dy2 = dy
+      for (const blockBB of surroundingBBs) {
+        dy1 = blockBB.computeOffsetY(BB_XZ, dy1)
+        dy2 = blockBB.computeOffsetY(BB2, dy2)
+      }
+      BB1.offset(0, dy1, 0)
+      BB2.offset(0, dy2, 0)
+
+      let dx1 = oldVelX
+      let dx2 = oldVelX
+      for (const blockBB of surroundingBBs) {
+        dx1 = blockBB.computeOffsetX(BB1, dx1)
+        dx2 = blockBB.computeOffsetX(BB2, dx2)
+      }
+      BB1.offset(dx1, 0, 0)
+      BB2.offset(dx2, 0, 0)
+
+      let dz1 = oldVelZ
+      let dz2 = oldVelZ
+      for (const blockBB of surroundingBBs) {
+        dz1 = blockBB.computeOffsetZ(BB1, dz1)
+        dz2 = blockBB.computeOffsetZ(BB2, dz2)
+      }
+      BB1.offset(0, 0, dz1)
+      BB2.offset(0, 0, dz2)
+
+      const norm1 = dx1 * dx1 + dz1 * dz1
+      const norm2 = dx2 * dx2 + dz2 * dz2
+
+      if (norm1 > norm2) {
+        dx = dx1
+        dy = -dy1
+        dz = dz1
+        playerBB = BB1
+      } else {
+        dx = dx2
+        dy = -dy2
+        dz = dz2
+        playerBB = BB2
+      }
+
+      for (const blockBB of surroundingBBs) {
+        dy = blockBB.computeOffsetY(playerBB, dy)
+      }
+      playerBB.offset(0, dy, 0)
+
+      if (oldVelXCol * oldVelXCol + oldVelZCol * oldVelZCol >= dx * dx + dz * dz) {
+        dx = oldVelXCol
+        dy = oldVelYCol
+        dz = oldVelZCol
+        playerBB = oldBBCol
+      }
+    }
+
+    // Update flags
+    setPositionToBB(playerBB, pos)
+    entity.isCollidedHorizontally = dx !== oldVelX || dz !== oldVelZ
+    entity.isCollidedVertically = dy !== oldVelY
+    entity.onGround = entity.isCollidedVertically && oldVelY < 0
+
+    const blockAtFeet = world.getBlock(pos.offset(0, -0.2, 0))
+
+    if (dx !== oldVelX) vel.x = 0
+    if (dz !== oldVelZ) vel.z = 0
+    if (dy !== oldVelY) {
+      if (blockAtFeet && blockAtFeet.type === slimeBlockId && !entity.control.sneak) {
+        vel.y = -vel.y
+      } else {
+        vel.y = 0
+      }
+    }
+
+    // Finally, apply block collisions (web, soulsand...)
+    playerBB.contract(0.001, 0.001, 0.001)
+    const cursor = new Vec3(0, 0, 0)
+    for (cursor.y = Math.floor(playerBB.minY); cursor.y <= Math.floor(playerBB.maxY); cursor.y++) {
+      for (cursor.z = Math.floor(playerBB.minZ); cursor.z <= Math.floor(playerBB.maxZ); cursor.z++) {
+        for (cursor.x = Math.floor(playerBB.minX); cursor.x <= Math.floor(playerBB.maxX); cursor.x++) {
+          const block = world.getBlock(cursor)
+          if (block) {
+            if (block.type === soulsandId) {
+              vel.x *= 0.4
+              vel.z *= 0.4
+            } else if (block.type === webId) {
+              entity.isInWeb = true
+            }
+          }
+        }
+      }
+    }
+  }
+
+  function applyHeading (entity, strafe, forward, multiplier) {
+    let speed = Math.sqrt(strafe * strafe + forward * forward)
+    if (speed < 0.01) return new Vec3(0, 0, 0)
+
+    speed = multiplier / Math.max(speed, 1)
+
+    strafe *= speed
+    forward *= speed
+
+    const yaw = Math.PI - entity.yaw
+    const sin = Math.sin(yaw)
+    const cos = Math.cos(yaw)
+
+    const vel = entity.vel
+    vel.x += strafe * cos - forward * sin
+    vel.z += forward * cos + strafe * sin
+  }
+
+  function isOnLadder (world, pos) {
+    const block = world.getBlock(pos)
+    return (block && (block.type === ladderId || block.type === vineId))
+  }
+
+  function moveEntityWithHeading (entity, world, strafe, forward) {
+    const vel = entity.vel
+    const pos = entity.pos
+
+    if (!entity.isInWater && !entity.isInLava) {
+      // Normal movement
+      let acceleration = 0.02 // airborne
+      let inertia = 0.91
+      const blockUnder = world.getBlock(pos.offset(0, -1, 0))
+      if (entity.onGround && blockUnder) {
+        inertia = (blockSlipperiness[blockUnder.type] || 0.6) * 0.91
+        acceleration = 0.1 * (0.1627714 / (inertia * inertia * inertia))
+      }
+      if (entity.control.sprint) acceleration *= physics.sprintSpeed
+
+      applyHeading(entity, strafe, forward, acceleration)
+
+      if (isOnLadder(world, pos)) {
+        const ladderMaxSpeed = 0.15
+        vel.x = math.clamp(-ladderMaxSpeed, vel.x, ladderMaxSpeed)
+        vel.z = math.clamp(-ladderMaxSpeed, vel.z, ladderMaxSpeed)
+        vel.y = Math.max(vel.y, entity.control.sneak ? 0 : -ladderMaxSpeed)
+      }
+
+      moveEntity(entity, world, vel.x, vel.y, vel.z)
+
+      if (entity.isCollidedHorizontally && isOnLadder(world, pos)) {
+        vel.y = 0.2 // climb ladder
+      }
+
+      // Apply friction and gravity
+      vel.y -= physics.gravity
+      vel.y *= 0.98 // air drag
+      vel.x *= inertia
+      vel.z *= inertia
+    } else {
+      // Water / Lava movement
+      const acceleration = 0.02
+      const inertia = entity.isInWater ? 0.8 : 0.5
+      // TODO: depth strider enchantement (for water)
+      applyHeading(entity, strafe, forward, acceleration)
+      moveEntity(entity, world, vel.x, vel.y, vel.z)
+      vel.y *= inertia
+      vel.y -= 0.02
+      vel.x *= inertia
+      vel.z *= inertia
+
+      if (entity.isCollidedHorizontally) { // TODO: && isOffsetPositionInLiquid
+        vel.y = 0.3 // jump out of liquid
+      }
+    }
+  }
+
+  function isMaterialInBB (world, queryBB, type) {
+    const cursor = new Vec3(0, 0, 0)
+    for (cursor.y = Math.floor(queryBB.minY); cursor.y <= Math.floor(queryBB.maxY); cursor.y++) {
+      for (cursor.z = Math.floor(queryBB.minZ); cursor.z <= Math.floor(queryBB.maxZ); cursor.z++) {
+        for (cursor.x = Math.floor(queryBB.minX); cursor.x <= Math.floor(queryBB.maxX); cursor.x++) {
+          const block = world.getBlock(cursor)
+          if (block && block.type === type) return true
+        }
+      }
+    }
+    return false
+  }
+
+  function getLiquidHeightPcent (block) {
+    return (getRenderedDepth(block) + 1) / 9
+  }
+
+  function getRenderedDepth (block) {
+    if (!block || block.type !== waterId) return -1
+    const meta = block.metadata
+    return meta >= 8 ? 0 : meta
+  }
+
+  function getFlow (world, block) {
+    const curlevel = getRenderedDepth(block)
+    const flow = new Vec3(0, 0, 0)
+    for (const [dx, dz] of [[0, 1], [-1, 0], [0, -1], [1, 0]]) {
+      const adjBlock = world.getBlock(block.position.offset(dx, 0, dz))
+      const adjLevel = getRenderedDepth(adjBlock)
+      if (adjLevel < 0) {
+        if (adjBlock.boundingBox !== 'empty') {
+          const adjLevel = getRenderedDepth(world.getBlock(block.position.offset(dx, -1, dz)))
+          if (adjLevel >= 0) {
+            const f = adjLevel - (curlevel - 8)
+            flow.x += dx * f
+            flow.z += dz * f
+          }
+        }
+      } else {
+        const f = adjLevel - curlevel
+        flow.x += dx * f
+        flow.z += dz * f
+      }
+    }
+
+    if (block.metadata >= 8) {
+      for (const [dx, dz] of [[0, 1], [-1, 0], [0, -1], [1, 0]]) {
+        const adjBlock = world.getBlock(block.position.offset(dx, 0, dz))
+        const adjUpBlock = world.getBlock(block.position.offset(dx, 1, dz))
+        if (adjBlock.boundingBox !== 'empty' || adjUpBlock.boundingBox !== 'empty') {
+          flow.normalize().translate(0, -6, 0)
+        }
+      }
+    }
+
+    return flow.normalize()
+  }
+
+  function isInWaterApplyCurrent (world, bb, vel) {
+    let isInWater = false
+    const acceleration = new Vec3(0, 0, 0)
+    const cursor = new Vec3(0, 0, 0)
+    for (cursor.y = Math.floor(bb.minY); cursor.y <= Math.floor(bb.maxY); cursor.y++) {
+      for (cursor.z = Math.floor(bb.minZ); cursor.z <= Math.floor(bb.maxZ); cursor.z++) {
+        for (cursor.x = Math.floor(bb.minX); cursor.x <= Math.floor(bb.maxX); cursor.x++) {
+          const block = world.getBlock(cursor)
+          if (block && block.type === waterId) {
+            const waterLevel = cursor.y + 1 - getLiquidHeightPcent(block)
+            if (Math.ceil(bb.maxY) >= waterLevel) {
+              isInWater = true
+              const flow = getFlow(world, block)
+              acceleration.add(flow)
+            }
+          }
+        }
+      }
+    }
+
+    const len = acceleration.norm()
+    if (len > 0) {
+      vel.x += acceleration.x / len * 0.014
+      vel.y += acceleration.y / len * 0.014
+      vel.z += acceleration.z / len * 0.014
+    }
+    return isInWater
+  }
+
+  physics.simulatePlayer = (entity, world) => {
+    const vel = entity.vel
+    const pos = entity.pos
+
+    const waterBB = getPlayerBB(pos).contract(0.001, 0.401, 0.001)
+    const lavaBB = getPlayerBB(pos).contract(0.1, 0.4, 0.1)
+
+    entity.isInWater = isInWaterApplyCurrent(world, waterBB, vel)
+    entity.isInLava = isMaterialInBB(world, lavaBB, lavaId)
+
+    // Reset velocity component if it falls under the threshold
+    if (Math.abs(vel.x) < physics.negligeableVelocity) vel.x = 0
+    if (Math.abs(vel.y) < physics.negligeableVelocity) vel.y = 0
+    if (Math.abs(vel.z) < physics.negligeableVelocity) vel.z = 0
+
+    // Handle inputs
+    if (entity.control.jump || entity.jumpQueued) {
+      if (entity.jumpTicks > 0) entity.jumpTicks--
+      if (entity.isInWater || entity.isInLava) {
+        vel.y += 0.04
+      } else if (entity.onGround && entity.jumpTicks === 0) {
+        vel.y = 0.42
+        // TODO: jump potion effect
+        if (entity.control.sprint) {
+          const yaw = Math.PI - entity.yaw
+          vel.x -= Math.sin(yaw) * 0.2
+          vel.z += Math.cos(yaw) * 0.2
+        }
+        entity.jumpTicks = 10 // activate autojump cooldown (0.5s)
+      }
+    } else {
+      entity.jumpTicks = 0 // reset autojump cooldown
+    }
+    entity.jumpQueued = false
+
+    let strafe = (entity.control.right - entity.control.left) * 0.98
+    let forward = (entity.control.forward - entity.control.back) * 0.98
+
+    if (entity.control.sneak) {
+      strafe *= 0.3
+      forward *= 0.3
+    }
+
+    moveEntityWithHeading(entity, world, strafe, forward)
+
+    return entity
+  }
+
+  return physics
+}
+
+class PlayerState {
+  constructor (bot, control) {
+    // Input / Outputs
+    this.pos = bot.entity.position.clone()
+    this.vel = bot.entity.velocity.clone()
+    this.onGround = bot.entity.onGround
+    this.isInWater = bot.entity.isInWater
+    this.isInLava = bot.entity.isInLava
+    this.isInWeb = bot.entity.isInWeb
+    this.isCollidedHorizontally = bot.entity.isCollidedHorizontally
+    this.isCollidedVertically = bot.entity.isCollidedVertically
+    this.jumpTicks = bot.jumpTicks
+    this.jumpQueued = bot.jumpQueued
+
+    // Input only (not modified)
+    this.yaw = bot.entity.yaw
+    this.control = control
+  }
+
+  apply (bot) {
+    bot.entity.position = this.pos
+    bot.entity.velocity = this.vel
+    bot.entity.onGround = this.onGround
+    bot.entity.isInWater = this.isInWater
+    bot.entity.isInLava = this.isInLava
+    bot.entity.isInWeb = this.isInWeb
+    bot.entity.isCollidedHorizontally = this.isCollidedHorizontally
+    bot.entity.isCollidedVertically = this.isCollidedVertically
+    bot.jumpTicks = this.jumpTicks
+    bot.jumpQueued = this.jumpQueued
+  }
+}
+
+module.exports = { Physics, PlayerState }
